Upgraded to Forge 1.8.8!

+++ src/main/java/gigaherz/elementsofpower/blocks/BlockDust.java
@@ -20,7 +20,6 @@ import java.util.Random;
 
 public class BlockDust extends Block
 {
-
     public static final PropertyInteger DENSITY = PropertyInteger.create("density", 1, 16);
 
     public BlockDust()
@@ -41,7 +40,7 @@ public class BlockDust extends Block
         IBlockState state = world.getBlockState(pos);
         if (state.getBlock() != this)
             return 16;
-        return (Integer) state.getValue(DENSITY);
+        return state.getValue(DENSITY);
     }
 
     @Override
@@ -74,7 +73,7 @@ public class BlockDust extends Block
     @Override
     public void updateTick(World worldIn, BlockPos pos, IBlockState state, Random rand)
     {
-        int density = (Integer) state.getValue(DENSITY) - 1;
+        int density = state.getValue(DENSITY) - 1;
         int maxGive = (int) Math.sqrt(density);
 
         for (EnumFacing f : EnumFacing.VALUES)
@@ -103,7 +102,7 @@ public class BlockDust extends Block
             {
                 if (density > maxGive)
                 {
-                    int od = (Integer) neighbour.getValue(DENSITY);
+                    int od = neighbour.getValue(DENSITY);
                     if (od < 16)
                     {
                         int d = rand.nextInt(Math.min(16 - od, maxGive));
@@ -152,7 +151,7 @@ public class BlockDust extends Block
     @Override
     public int getMetaFromState(IBlockState state)
     {
-        return 16 - (Integer) state.getValue(DENSITY);
+        return 16 - state.getValue(DENSITY);
     }
 
     @Override
